+++ bindings/dart/keri/keri_macos/lib/keri_macos.dart
@@ -563,4 +563,19 @@ class KeriMacos extends KeriPlatformInterface {
     return await api.finalizeQuery(
         identifier: identifier, queryEvent: queryEvent, signature: signature);
   }
+
+  Future<bool> notifyWitnesses(
+      {required Identifier identifier, dynamic hint}) async {
+    return await api.notifyWitnesses(identifier: identifier);
+  }
+
+  Future<bool> broadcastReceipts(
+      {required Identifier identifier,
+        required List<Identifier> witnessList,
+        dynamic hint}) async {
+    return await api.broadcastReceipts(
+        identifier: identifier, witnessList: witnessList);
+  }
+
+
 }