+++ bindings/dart/keri/keri_macos/CHANGELOG.md
@@ -5,3 +5,7 @@ Initial release
 ## 0.0.2
 
 Aligned with `keri_platform_interface` 0.0.5
+
+## 0.0.3
+
+Added `notifyWitnesses` and `broadcastReceipts` functions